Refactor - Rewrite Dialog_Saturation, Rename To Dialog_Hue_Saturation And Update Filters Logic

+++ Anim-Paint/include/Filters.hpp
@@ -5,11 +5,11 @@ extern std::string palette_button_shader_source;
 extern std::string palette_colors_shader_source;
 extern std::string palette_values_shader_source;
 
-void set_rotation(sf::Image& image, float angle, bool set_smooth, sf::Color backgroundColor);
-void set_brightness(sf::Image& image, float value);
-void set_contrast(sf::Image& image, float value);
-void set_saturation(sf::Image& image, float value);
-void set_sepia(sf::Image& image, float value);
+void set_rotation(sf::Image& image, int angle, bool set_smooth, sf::Color backgroundColor);
+void set_brightness(sf::Image& image, int value);
+void set_contrast(sf::Image& image, int value);
+void set_saturation(sf::Image& image, int value);
+void set_sepia(sf::Image& image, int value);
 void set_outline(sf::Image& image, int width, sf::Color backgroundColor, sf::Color outlineColor);
 void set_resize(sf::Image& image, int width, int height, sf::Color backgroundColor);
 void set_chessboard(sf::Image& image, int tileCount, int transparency, sf::Color firstColor, sf::Color secondColor);
